+++ pkg/reexec/MAINTAINERS
@@ -1,0 +1,1 @@
+Michael Crosby <michael@docker.com> (@crosbymichael)
